requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.12.4
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pycodestyle==2.8.0
pyparsing==3.0.6
pytz==2021.1
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
uritemplate==4.1.1
urllib3==1.26.7
